Rename test files. Add  a test class for local disk provider with buit-in metaprovider.
