+++ src/Tests/SenseNet.BlobStorage.IntegrationTests/BuiltInFsTests.cs
@@ -1,108 +1,110 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SenseNet.BlobStorage.IntegrationTests.Implementations;
-using SenseNet.MsSqlFsBlobProvider;
+using SenseNet.ContentRepository;
+using SenseNet.ContentRepository.Storage;
+using SenseNet.ContentRepository.Storage.Data;
+using SenseNet.ContentRepository.Storage.Data.SqlClient;
 
 namespace SenseNet.BlobStorage.IntegrationTests
 {
     [TestClass]
-    public class LocalDiskBlobProviderTests : BlobStorageIntegrationTests
+    public class BuiltInFsTests : BlobStorageIntegrationTests
     {
         protected override string DatabaseName => "sn7blobtests_builtinfs";
         protected override bool SqlFsEnabled => true;
         protected override bool SqlFsUsed => false;
-        protected override Type ExpectedExternalBlobProviderType => typeof(LocalDiskBlobProvider);
-        protected override Type ExpectedMetadataProviderType => typeof(SqlFileStreamBlobMetaDataProvider);
-        protected override Type ExpectedBlobProviderDataType => typeof(LocalDiskBlobProvider.LocalDiskBlobProviderData);
+        protected override Type ExpectedExternalBlobProviderType => null;
+        protected override Type ExpectedMetadataProviderType => typeof(MsSqlBlobMetaDataProvider);
+        protected override Type ExpectedBlobProviderDataType => typeof(BuiltinBlobProviderData);
 
         protected internal override void ConfigureMinimumSizeForFileStreamInBytes(int newValue, out int oldValue)
         {
-            oldValue = Configuration.BlobStorage.MinimumSizeForBlobProviderInBytes;
-            Configuration.BlobStorage.MinimumSizeForBlobProviderInBytes = newValue;
+            // do nothing
+            oldValue = 0;
         }
 
         [ClassCleanup]
         public static void CleanupClass()
         {
-            TearDown(typeof(LocalDiskBlobProviderTests));
+            TearDown(typeof(BuiltInFsTests));
         }
 
+        /* ==================================================== Test cases */
 
         [TestMethod]
-        public void Blob_LocalDisk_01_CreateFileSmall()
+        public void Blob_BuiltInFS_01_CreateFileSmall()
         {
             TestCase01_CreateFileSmall();
         }
         [TestMethod]
-        public void Blob_LocalDisk_02_CreateFileBig()
+        public void Blob_BuiltInFS_02_CreateFileBig()
         {
             TestCase02_CreateFileBig();
         }
 
         [TestMethod]
-        public void Blob_LocalDisk_03_UpdateFileSmallSmall()
+        public void Blob_BuiltInFS_03_UpdateFileSmallSmall()
         {
             TestCase03_UpdateFileSmallSmall();
         }
         [TestMethod]
-        public void Blob_LocalDisk_04_UpdateFileSmallBig()
+        public void Blob_BuiltInFS_04_UpdateFileSmallBig()
         {
             TestCase04_UpdateFileSmallBig();
         }
         [TestMethod]
-        public void Blob_LocalDisk_05_UpdateFileBigSmall()
+        public void Blob_BuiltInFS_05_UpdateFileBigSmall()
         {
             TestCase05_UpdateFileBigSmall();
         }
         [TestMethod]
-        public void Blob_LocalDisk_06_UpdateFileBigBig()
+        public void Blob_BuiltInFS_06_UpdateFileBigBig()
         {
             TestCase06_UpdateFileBigBig();
         }
 
         [TestMethod]
-        public void Blob_LocalDisk_07_WriteChunksSmall()
+        public void Blob_BuiltInFS_07_WriteChunksSmall()
         {
             TestCase07_WriteChunksSmall();
         }
         [TestMethod]
-        public void Blob_LocalDisk_08_WriteChunksBig()
+        public void Blob_BuiltInFS_08_WriteChunksBig()
         {
             TestCase08_WriteChunksBig();
         }
 
         [TestMethod]
-        public void Blob_LocalDisk_09_DeleteBinaryPropertySmall()
+        public void Blob_BuiltInFs_09_DeleteBinaryPropertySmall()
         {
             TestCase09_DeleteBinaryPropertySmall();
         }
         [TestMethod]
-        public void Blob_LocalDisk_10_DeleteBinaryPropertyBig()
+        public void Blob_BuiltInFs_10_DeleteBinaryPropertyBig()
         {
             TestCase10_DeleteBinaryPropertyBig();
         }
 
         [TestMethod]
-        public void Blob_LocalDisk_11_CopyfileRowSmall()
+        public void Blob_BuiltInFs_11_CopyfileRowSmall()
         {
             TestCase11_CopyfileRowSmall();
         }
         [TestMethod]
-        public void Blob_LocalDisk_12_CopyfileRowBig()
+        public void Blob_BuiltInFs_12_CopyfileRowBig()
         {
             TestCase12_CopyfileRowBig();
         }
 
         [TestMethod]
-        public void Blob_LocalDisk_13_BinaryCacheEntitySmall()
+        public void Blob_BuiltInFs_13_BinaryCacheEntitySmall()
         {
             TestCase13_BinaryCacheEntitySmall();
         }
         [TestMethod]
-        public void Blob_LocalDisk_14_BinaryCacheEntityBig()
+        public void Blob_BuiltInFs_14_BinaryCacheEntityBig()
         {
             TestCase14_BinaryCacheEntityBig();
         }
-
     }
 }
